chiper_Totoro/requirements.txt
django-sslserver==0.20
django-ratelimit==2.0.0
cryptography==2.8
Django==2.2
psycopg2==2.8.4
psycopg2-binary==2.8.4
PyJWT==1.7.1
python-dotenv==0.10.1
social-auth-app-django==3.1.0
django-widget-tweaks==1.4.8